coverity#736885 Unintended sign extension
and coverity#736886 Unintended sign extension

promote from unsigned short to size_t immediately rather than implicitly to int
and then have the int result widened to size_t

+++ lotuswordpro/source/filter/lwptablelayout.hxx
@@ -102,7 +102,7 @@ public:
     {
         if (nRow >= m_nRows || nCol >= m_nCols)
             return;
-        m_WordProCellsMap[nRow * m_nCols + nCol] = pCell;
+        m_WordProCellsMap[static_cast<size_t>(nRow) * m_nCols + nCol] = pCell;
     };
 protected:
     void Read() SAL_OVERRIDE;
